DO NOT SUBMIT: Temporary workaround to Authorization Server Metadata missing required fields.
[1] currently misses the required fields [2-3], which are both required per [4].
The OpenID4VCI library's Issuer.kt lines 343 [5] and 349 contain assertions
checking that the signature algorithm (ES256) that we have specified for the attestation
is listed there. To be able to make progress while the PID Provider is not yet reporting
this metadata, this commit is temporarily adding logic to pretend that both fields contain
ES256 when the fields are otherwise absent.

[1] https://demo.pid-provider.bundesdruckerei.de/.well-known/oauth-authorization-server
[2] client_attestation_signing_alg_values_supported
[3] client_attestation_pop_signing_alg_values_supported
[4] https://www.ietf.org/archive/id/draft-ietf-oauth-attestation-based-client-auth-07.html#name-authorization-server-metada
[5] https://github.com/german-first-iteration/eudi-lib-jvm-openid4vci-kt/blob/c93218585cae8ec10c5df5120c3b4b4a10954289/src/main/kotlin/eu/europa/ec/eudi/openid4vci/Issuer.kt#L343